requirements.txt
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
pytesseract==0.2.6
